+++ lib/domain/general_domain/entities/main_plot_data_point_entity.dart
@@ -1,6 +1,6 @@
 class MainPlotDataPointEntity {
   final String semester;
-  final int ects;
+  final double ects;
   final double averageGrade;
 
   MainPlotDataPointEntity(
